+++ Sources/VCore/Views/UIKit/Infinite Scrolling UI Views/Infinite Scrolling UI Collection View/InfiniteScrollingUICollectionView.swift
@@ -184,7 +184,7 @@ open class InfiniteScrollingUICollectionView: UICollectionView {
         }
         
         (collectionViewLayout as? UICollectionViewFlowLayout)?.footerReferenceSize = CGSize(
-            width: bounds.width,
+            width: bounds.size.width,
             height: ActivityIndicatorUIModel.height
         )
     }